move helpers into separate extensions
